+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.1
+
+removed registerEffects and mapActionToMap functions
+exposed effect function
+
 ## 1.0.0
 
 Framework agnostic state management tool without ceremonies and boilerplates